reorganized the application entity UI to move common info into the top (always visible) section
