+++ manager/ui/war/src/main/java/io/apiman/manager/ui/client/local/site/app-contracts.html
@@ -52,28 +52,66 @@
         </div>
       </div>
       <!-- Entity Summary Row -->
-      <div class="row">
-        <div class="col-md-12 apiman-entity-breadcrumb">
-          <div class="title container-fluid">
-            <i class="breadcrumb-icon fa fa-gears"></i>
-            <div class="entity emphasis">
-              <a data-field="application" href="app-overview.html" data-role="dummy">gadget-server</a>
-            </div>
-            <div class="versions">
-              <div class="btn-group apiman-entity-action" data-role="dummy" data-field="versions">
-                <button type="button" class="btn btn-default dropdown-toggle" data-toggle="dropdown">
-                  Version: 3.0.1 <span class="caret"></span>
-                </button>
-                <ul class="dropdown-menu" role="menu">
-                  <li><a href="#">3.0</a></li>
-                  <li><a href="#">2.1.1</a></li>
-                  <li><a href="#">2.1</a></li>
-                  <li><a href="#">2.0.1</a></li>
-                  <li><a href="#">2.0</a></li>
-                  <li><a href="#">1.0</a></li>
-                </ul>
+      <div class="apiman-entity-summary">
+        <div class="row apiman-entity-breadcrumb">
+          <div class="col-md-12">
+            <div class="title container-fluid">
+              <i class="breadcrumb-icon fa fa-gears"></i>
+              <div class="entity emphasis">
+                <a data-field="application" href="app-overview.html" data-role="dummy">gadget-server</a>
               </div>
-              <a data-permission="appEdit" href="new-appversion.html" class="btn btn-primary apiman-entity-action" data-field="toNewAppVersion" data-i18n-key="new-version">New Version</a>
+              <div class="versions">
+                <div class="btn-group apiman-entity-action" data-role="dummy" data-field="versions">
+                  <button type="button" class="btn btn-default dropdown-toggle" data-toggle="dropdown">
+                    Version: 3.0.1 <span class="caret"></span>
+                  </button>
+                  <ul class="dropdown-menu" role="menu">
+                    <li><a href="#">3.0</a></li>
+                    <li><a href="#">2.1.1</a></li>
+                    <li><a href="#">2.1</a></li>
+                    <li><a href="#">2.0.1</a></li>
+                    <li><a href="#">2.0</a></li>
+                    <li><a href="#">1.0</a></li>
+                  </ul>
+                </div>
+                <a data-permission="appEdit" href="new-appversion.html" class="btn btn-primary apiman-entity-action" data-field="toNewAppVersion" data-i18n-key="new-version">New Version</a>
+              </div>
+            </div>
+            <hr />
+          </div>
+        </div>
+        <div class="row apiman-entity-metadata">
+          <div class="col-md-7" style="margin-bottom: 8px">
+            <!-- Service Summary -->
+            <div class="description" data-field="description" data-role="dummy">This is a project that builds on the Apache Shindig as the open social gadget containers.</div>
+            <div style="padding: 8px">
+              <div class="entity-info-with-icon">
+                <i class="fa fa-clock-o fa-fw"></i>
+                <span class="apiman-label-faded" data-i18n-key="created-on">Created on</span>
+                <span data-field="createdOn" data-role="dummy" >Feb 3, 2011</span>
+              </div>
+              <div class="entity-info-with-icon">
+                <i class="fa fa-user fa-fw"></i>
+                <span class="apiman-label-faded" data-i18n-key="created-by">Created by</span>
+                <span><a href="user-orgs.html" data-field="createdBy" data-role="dummy">Clark Kent</a></span>
+              </div>
+            </div>
+            <div class="entity-info-with-icon">
+              <span data-i18n-key="status-label">Status:</span>
+              <span class="apiman-label apiman-label-warning" data-field="status" data-role="dummy">Created</span>
+            </div>
+          </div>
+          <div class="col-md-5" data-permission="appEdit">
+            <div>
+              <div data-status="Created,Ready"><a data-i18n-key="ttdo-consume-services" data-field="ttd_toConsumeServices" href="consumer-services.html">Search for Services to consume</a></div>
+              <div data-status="Created,Ready"><a data-i18n-key="ttdo-new-contract" data-field="ttd_toNewContract" href="new-contract.html">Create a new Service Contract for this Application</a></div>
+              <div><a data-i18n-key="ttdo-new-version" data-field="ttd_toNewVersion" href="new-appversion.html">Create a new version of this Application (New Version)</a></div>
+            </div>
+            <!-- The Publish Action -->
+            <div class="apiman-divider-40"></div>
+            <div class="">
+              <button class="btn btn-primary" data-field="registerButton" data-i18n-key="register" placeholder="Registering..." data-icon="fa-cog">Register</button>
+              <button class="btn btn-warning" data-field="unregisterButton" data-i18n-key="unregister" placeholder="Unregistering..." data-icon="fa-cog">Unregister</button>
             </div>
           </div>
         </div>
@@ -85,10 +123,10 @@
         <div class="col-md-2 apiman-entity-nav">
           <ul class="nav nav-pills nav-stacked">
             <li class="first"><a data-role="dummy">&nbsp;</a></li>
-            <li class="north"><a data-field="toAppOverview" href="app-overview.html" data-i18n-key="overview">Overview</a></li>
+            <li class=""><a data-field="toAppOverview" href="app-overview.html" data-i18n-key="overview">Overview</a></li>
             <li class="active"><a data-field="toAppContracts" href="app-contracts.html" data-i18n-key="contracts">Contracts</a></li>
-            <li class="south"><a data-field="toAppApis" href="app-apis.html" data-i18n-key="apis">APIs</a></li>
             <li class=""><a data-field="toAppPolicies" href="app-policies.html" data-i18n-key="policies">Policies</a></li>
+            <li data-status="Registered"><a data-field="toAppApis" href="app-apis.html" data-i18n-key="apis">APIs</a></li>
             <li><a data-field="toAppActivity" href="app-activity.html" data-i18n-key="activity">Activity</a></li>
             <li class="last"><a data-role="dummy">&nbsp;</a></li>
           </ul>
